requirements.txt
torch
torchvision
torchaudio
timm
requests
open3d
opencv-python
Pillow
